New compressing framework
In “fib.algorithms/nConvertToFib/nImage” a new compressing framwork for
images was implemented.
The normal method (for the methods / strategies
@see fib.algorithms/nConvertToFib/nImage) is to search (@see subfolder
nSearchForStructur) in the image for image structures
(@see subfolder nStructureData) and convert them into Fib objects
(@see subfolder nStructureData/nConvertToFib).
To adapt the found image structures operators
(@see subfolder nStructureData/nOperators/ )on them can be used.
The converter (@see add_ons/converter/ ) wher adapted acordingly.
And the new converter convertFromImageStructure was added, so that created
image structures can be converted to images (to analyse the process).

This is the first working (alpha) version of the framework.
Some more things are planned to be imlemented and some code review is still
to be done.