method assignValues() added
The method assignValues() was added for all Fib elements.
With this method values from one Fib element can be assigned to a
Fib element of the same type.
Test for the method were also implemented and run.

There was also a bugfix for generator of Fib objects implemented.
If rational numbers are generated, they now have a finit number of
digits in the binary representation (you can multiply them with two to get
an integer).