Choose Fib object dialog + FEATURE_INSERT_OBJECT_IN_ELEMENT implemented
Choose Fib object dialog

A new subwindow of the Fib creator GUI.
In this subwindow Fib objects can be displayed and selected.
It is for giving easy access to Fib objects.

FEATURE_INSERT_OBJECT_IN_ELEMENT implemented

With this feature a new version of insertObjectInElement() is implemented.
If the insert position is a list element, the insert object method will
now add the to insert Fib object into the list element on the position,
instead of creating a new list element.
Also if the "first" parameter (if true: the inserted object will be the
first subobject) is false, the inserted object will be inserted as the
last subobject of the branch element.

Some more minor bugfixes and changes