vector widget added
A new vector widget implemented, wich will show the content of a vector.
Some vector elements (values) can be manipulated to change the Fib object.

Grapical items can now show some options to change them, if the mouse is
above them.

A graphical item to show external objects added. It shows the input vector
as its option, if the mouse is above it.

Also some test Fib objects added, to play with the application.

Some minor bugfixes and changes.