analyse loaded (sub) objects
Objects loaded in the Fib object select dialog, will be analysed.
Feature: FEATURE_FIB_OBJECT_INFO_IN_VAR_TYPE_FROM_TEXT_REGEX
(regex needed -> just works with new g++ compilers)

The optional information will be scanned for information, which input
variable is of which type (e.g. input variable for the first point position
in dimension 1), this information will be stored.
The information is also used, if a Fib object is integrated into an other
Fib object (/node), so that for example the Fib object will be inserted at
the actual (mouse) location.

- some bugfixes
- code review